requirements.txt
# Multi-Agent Interview Coach
openai>=1.12.0
python-dotenv>=1.0.0
